Refactored JokePresentation to separate viewmodel
- separate view model for jokePresentation
- refactored corresponding navigation and passed variables
